Optimize by run single list --format=xml
The existing code is very slow with more than a couple of commands (especially so on a Docker for Mac osxfs host volume mount).

The symfony/console library supports fetching a single combined xml of all commands, so using this reduces the number of console command runs to one, massively speeding it up.