Start migration from Travis-CI to Github CI
Porting jobs matrix by the two build environs (composer /--prefer-lowest)
and PHP versions from 7.0 to 7.4 (and 8.0 + nightly w/ allow-failure).

Straight forward port from .travis.yml with the help of the
shivammathur/setup-php and ktomk/run-travis-yml Github Actions.

Additionally adding PHP 8.0 and nightly (8.1 @ shivammathur/setup-php).

Different is the Ubuntu version: Upstream was on Xenial (for base-line)
and on Trusty (for PHP 8.0 PR). Picking Xenial (Trusty is n/a on
Github).

Rationale is that the project is still running on travis-ci.org which
is shutting down. And: PHP 8.0 support (different issue).

PHP 8.0/8.1 fail by default, as not the PHP 8.0 PR nor already merged.

Further: bamarni/symfony-console-autocomplate is a dependency in magerun
and we have interest in a working CI and PHP 8.0 compatibility upstream.