fix symfony-autocomplete composer vendor
bin/symfony-autocomplete did not honor COMPOSER_VENDOR_DIR and treated it
as if it is the default value "vendor" only.

however it is part of the composer install protocol and required for
the bin-stubs autoloading.  it should not be hard-encoded to the default
value only.

fix is to use COMPOSER_VENDOR_DIR as absolute path, as relative path to
PWD and as relative path to bin/symfony-autocomplete given the first
attempt to try requiring from __DIR__.'/../../../autoload.php' failed.

NOTE: use of include_once instead of require_once is intended here, the EA
      inspection is misleading on it as _once was and still is intended.

this is similar for "acme" in tests, but with the difference that during
build time we insist to know that COMPOSER_VENDOR_DIR is always relative
to the project root, YMMV.

report: #76